+++ versioninfo.h
@@ -1,5 +1,5 @@
 //generated resource header 
 #define GEN_LATEST_VERSION_STRING "1.0.0\0" 
-#define GEN_COMMIT_STRING "94d97e51\0" 
-#define GEN_COMMIT_COUNT_STRING "516\0" 
-#define GEN_VERSION_STRING "1.0.0.516\0" 
+#define GEN_COMMIT_STRING "2cad8343\0" 
+#define GEN_COMMIT_COUNT_STRING "517\0" 
+#define GEN_VERSION_STRING "1.0.0.517\0"